Use one script to do all CSS processing
Moved all the CSS processing code into one location,
generate.css.php.

This meant that using __FILE__ to determine the basename for
the output file was no longer valid. Instead, use
$_SERVER['SCRIPT_URI'], which refers to the original script, rather
than any scripts being require'd.

The main impetus for this was so I could set the sass path
explicitly for the script, rather than just calling "sass" (there
may be situations where the Apache user doesn't have sass on its
path). "sass" is now the default, but can be overridden in Apache
config with "SetEnv SASS /path/to/sass". The HACKING.md file
explains how.
