+++ site.conf
@@ -1,5 +1,11 @@
 Alias /crosswalk-website "/home/me/dev/crosswalk-website"
 
+<IfModule mod_env.c>
+  # uncomment this line if your sass script is on a path inaccessible
+  # to the Apache user
+  #SetEnv SASS /usr/local/bin/sass
+</IfModule>
+
 <DirectoryMatch "^\..*$">
   Require all denied
 </DirectoryMatch>
